fix logout toastify msg

+++ src/common/Logout.jsx
@@ -2,14 +2,17 @@ import React, { useEffect } from "react";
 import { toast } from "react-toastify";
 
 export default function Logout() {
+  const hasRun = useRef(false); // useRef doesn’t trigger re-renders
   useEffect(() => {
-    console.log("@@@@@@@@@@");
-    localStorage.clear();
-    toast.success("Your are Logout Successfully");
+    // console.log("@@@@@@@@@@");
+    if (!hasRun) {
+      localStorage.clear();
+      toast.success("Your are Logout Successfully");
 
-    setTimeout(() => {
-      window.location.href = "/";
-    }, 2000);
+      setTimeout(() => {
+        window.location.href = "/";
+      }, 2000);
+    }
   }, []);
   return null;
 }
